fix: refactor, moved blocks
