+++ frontend/src/components/blocks/educations-starting-block/educations-starting-block.tsx
@@ -1,5 +1,5 @@
 import DropCard from '@components/card/drop-card.component';
-import { CardsBlock } from '@components/cards-block/cards-block.component';
+import { CardsBlock } from '@components/blocks/cards-block/cards-block.component';
 import { EducationsStartingBlock as EducationsStartingBlockType } from '@interfaces/admin-data';
 import DateRangeIcon from '@mui/icons-material/DateRange';
 import LocationOnIcon from '@mui/icons-material/LocationOn';
@@ -16,6 +16,7 @@ export default function EducationsStartingBlock({ educationsStartingBlock }: Edu
   if (!educationsStartingBlock) return null;
   return (
     <CardsBlock<GetEducationEvents['courses']>
+      className="EducationsStartingBlock"
       title={'Utbildningar som snart börjar'}
       cards={educationsStartingBlock?.courses ?? []}
       cardRender={(course, index) => {
