--- frontend/src/components/saved-content-block/saved-content-block-empty.component.tsx
@@ -1,7 +1,0 @@
-export default function SavedContentBlockEmpty({ children }: { children: React.ReactNode }) {
-  return (
-    <div className="text-[1.7rem] leading-[2.5rem] desktop:text-[2.1rem] desktop:leading-[3rem] p-md grow text-center flex items-center">
-      {children}
-    </div>
-  );
-}
